requirements.txt
asgiref==3.5.2
bcrypt==3.2.2
cffi==1.15.0
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.5
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
yapf==0.32.0
